requirements.txt
# Dependencias básicas
requests
pymongo
pika

# Procesamiento de archivos
pypdf
# python-docx
# beautifulsoup4

# Dependencias para OCR de PDF
# dependencias necesarias en el sistema
# tesseract
# tesseract-lang
# poppler

pillow
pdf2image
pytesseract
numpy
easyocr  # OCR avanzado (opcional, requiere PyTorch)
torch
torchvision

# Procesamiento de texto y RAG
markdown
langchain
langchain-text-splitters
nltk
sentence-transformers
transformers

# Base de datos vectorial
qdrant-client

# Procesamiento en tiempo real
bytewax

# Utilidades
pydantic
loguru
uv
python-dotenv
ruff